+++ HISTORY.md
@@ -5,3 +5,4 @@
 
 [ version 2021.12.21 ]
 - change js_api parameter of 'create_window' from single only to multiple accesible
+- add target window attribute to 'BrowserAPI'